fix: pass more tests
